* added support for long and ulong to Bits utility * Bits enum tests now work for 64-bit enum types * added BitSet types * added ICache, LruCache implementations * starting on linked list table implementation * added open/close callbacks to DMMenu * improved MeshData bounds checking * MeshData support for different mesh topologies * added support for measuring in microseconds and ticks to Profiling.Time

+++ Assets/BeauUtil/Strings/Match/MatchRuleSet.cs
@@ -244,7 +244,7 @@ namespace BeauUtil
 
             if (m_EntryCount >= m_Entries.Length)
             {
-                Array.Resize(ref m_Entries, (int) Unsafe.AlignUp8((uint) m_Entries.Length + 1));
+                Array.Resize(ref m_Entries, Unsafe.AlignUp8(m_Entries.Length + 1));
             }
 
             m_Entries[m_EntryCount++] = newEntry;
